+++ src/webview/icons/index.ts
@@ -25,6 +25,7 @@ import eventSvg from "./event.svg";
 import userSvg from "./user.svg";
 import tagSvg from "./tag.svg";
 import externalLinkSvg from "./external-link.svg";
+import arrowsRotateSvg from "./arrows-rotate.svg";
 import notdefSvg from "./notdef.svg";
 
 export const icons = {
@@ -47,6 +48,7 @@ export const icons = {
   user: userSvg,
   tag: tagSvg,
   "external-link": externalLinkSvg,
+  "arrows-rotate": arrowsRotateSvg,
   notdef: notdefSvg,
 } as const;